Added confirmation dialog before a habit deletion

+++ src/routes/habitos/Habit.js
@@ -7,16 +7,21 @@ import { UserContext } from '../../contexts/UserContext';
 
 export default function Habit({name, habitDays, habitID, refreshList}) {
     const {token} = useContext(UserContext)
-    
+
+    function confirmDeletion(){
+        const confirmation = window.confirm(`Deseja realmente remover o hábito "${name}"?`);
+            if (confirmation) {
+                deleteHabit(habitID, token)
+                    .then(refreshList)
+                    .catch(()=>alert('Erro'));    
+            }
+    }
+
     return (
         <S.Habit>
             <span>{name}</span>
 
-            <GoTrashcan onClick={()=>{
-                deleteHabit(habitID, token)
-                .then(refreshList)
-                .catch(()=>alert('Erro'));
-            }}/>
+            <GoTrashcan onClick={()=>confirmDeletion()}/>
 
             <WeekCheckboxes 
             disabled={true}
